mastercam-pdm/backend/requirements.txt
fastapi
uvicorn[standard]
GitPython
requests
pydantic
cryptography
psutil
bcrypt
PyJWT